Speed up interactive generation
With a horrible, horrible eval.

A simple way to avoid this would be to be able to require the Thorfile
with some syntax I'm not aware of, or possibly rename the Thofile to
something that ends in "rb" and still works by default.

Another possible solution is to place the actual thor class in a
separate file in lib and require it from both the Thorfile and the
Guardfile.